adding profile pic finished
