+++ app/src/main/java/fistbumpstudios/fistbump/setUserName.java
@@ -3,21 +3,15 @@ package fistbumpstudios.fistbump;
 import android.Manifest;
 import android.content.Context;
 import android.content.Intent;
-import android.database.Cursor;
-import android.graphics.BitmapFactory;
-import android.net.Uri;
 import android.net.wifi.WifiInfo;
 import android.net.wifi.WifiManager;
 import android.os.Bundle;
-import android.provider.MediaStore;
 import android.support.v4.app.ActivityCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.Toast;
 
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 
@@ -49,17 +43,14 @@ public class setUserName extends AppCompatActivity {
 
             //FileOutputStream fos = new FileOutputStream("user_info.txt");
             fos.write((username + "\n").getBytes());
-            fos.write((getMAC() + "\n").getBytes());
-            fos.write(picturePath.getBytes());
+            fos.write((getMAC()).getBytes());
             fos.close();
 
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
-    public void verifyFields(View view){
+    public void verifyUsername(View view){
 
         EditText mEdit=(EditText)findViewById(R.id.username);
         String uname = mEdit.getText().toString();
@@ -69,45 +60,10 @@ public class setUserName extends AppCompatActivity {
                     Toast.LENGTH_SHORT).show();
             return;
         }
-
-        if(picturePath.equals("")){
-            Toast.makeText(this, "Select a profile picture!" + uname,
-                    Toast.LENGTH_SHORT).show();
-            return;
-
-        }
         makeVerifyFile(uname);
-        Intent intent = new Intent(this, tabbedMain.class);
+        Intent intent = new Intent(this, setProfilePic.class);
         startActivity(intent);
         finish();
     }
 
-    public void selectProfilePic(View view){
-        Intent i = new Intent(
-                Intent.ACTION_PICK,
-                android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
-
-        startActivityForResult(i, RESULT_LOAD_IMAGE);
-
-    }
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if (requestCode == RESULT_LOAD_IMAGE && resultCode == RESULT_OK && null != data) {
-            Uri selectedImage = data.getData();
-            String[] filePathColumn = {MediaStore.Images.Media.DATA};
-
-            Cursor cursor = getContentResolver().query(selectedImage,
-                    filePathColumn, null, null, null);
-            cursor.moveToFirst();
-
-            int columnIndex = cursor.getColumnIndex(filePathColumn[0]);
-            picturePath = cursor.getString(columnIndex);
-            cursor.close();
-
-            ImageView imageView = (ImageView) findViewById(R.id.profilePreview);
-            imageView.setImageBitmap(BitmapFactory.decodeFile(picturePath));
-        }
-    }
 }
