+++ app/src/main/res/layout/activity_set_profile_pic.xml
@@ -9,4 +9,43 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="fistbumpstudios.fistbump.setProfilePic">
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="Click on the image to set your profile picture"
+        android:id="@+id/textView5"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentStart="true" />
+
+    <ImageView
+        android:layout_marginTop="20dp"
+        android:layout_width="match_parent"
+        android:layout_height="130dp"
+        android:id="@+id/profilePreview"
+        android:onClick="selectProfilePic"
+        android:src="@drawable/placeholder"
+        android:layout_below="@+id/textView5"
+        android:layout_alignParentStart="true"
+        android:layout_above="@+id/nextButton" />
+
+    <Button
+        android:layout_marginTop="20dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="next"
+        android:id="@+id/nextButton"
+        android:onClick="verifyPic"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentEnd="true"
+        android:layout_marginBottom="183dp" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Your picture can be changed later in settings"
+        android:id="@+id/textView2"
+        android:layout_marginTop="68dp"
+        android:layout_alignTop="@+id/nextButton"
+        android:layout_centerHorizontal="true" />
 </RelativeLayout>
